pyproject.toml
[tool.poetry]
name = "cars-api"
version = "0.1.0"
description = ""
authors = ["Dennis Wainaina <denny.nganga@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.2"
alembic = "^1.13.1"
taskipy = "^1.12.2"
sqlalchemy = "^2.0.29"
uvicorn = "^0.29.0"
python-slugify = "^8.0.4"
click = "^8.1.7"
pydantic-settings = "^2.2.1"
bcrypt = "^4.1.3"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.1"
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
load_initial_data = { cmd = "python -m scripts.load_initial_data", help = "Load initial car data to the DB" }
